requirements.txt
discord.py
python-dotenv
guidance
pymongo
google-generativeai
google-cloud-aiplatform